refactor: Переносит компонент формы в shared слой
