+++ src/pages/sign-up-page/ui/sign-up-page.ts
@@ -1,6 +1,6 @@
 import { AppRoutes } from "shared/constants";
 import { Block } from "shared/constructors";
-import { Form } from "shared/form";
+import { Form } from "shared/ui";
 import {
   validateEmail,
   validateLogin,
